+++ OpenNos.DAL.EF.MySQL/DB/ShopItem.cs
@@ -15,7 +15,7 @@ namespace OpenNos.DAL.EF.MySQL.DB
     public partial class ShopItem
     {
         public int ShopItemId { get; set; }
-        public short Type { get; set; }
+        public byte Type { get; set; }
         public short Slot { get; set; }
         public short ItemVNum { get; set; }
         public byte Upgrade { get; set; }
