+++ OpenNos.Data/ItemDTO.cs
@@ -74,7 +74,7 @@ namespace OpenNos.Data
         public byte ReputationMinimum { get; set; }
         public byte FairyMaximumLevel { get; set; }
         public byte MaximumAmmo { get; set; }
-        public short BasicUpgrade { get; set; }
+        public byte BasicUpgrade { get; set; }
         public short Color { get; set; }
         public long ItemValidTime { get; set; }
         public bool IsPearl { get; set; }
